requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
requests
numpy
faiss-cpu
pytest
